cleanup the app
