+++ emulator/app/src/tests/mctp_util/common.rs
@@ -142,7 +142,7 @@ impl MctpUtil {
                 I3cControllerState::Start => {
                     // Add some delay before sending the first packet.
                     // The MCU might need some time to boot up and be ready to receive the request.
-                    std::thread::sleep(std::time::Duration::from_secs(5));
+                    std::thread::sleep(std::time::Duration::from_secs(10));
                     i3c_state = I3cControllerState::SendPrivateWrite;
                 }
