+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/cert_slot_mgr/mod.rs
@@ -1,6 +1,6 @@
 // Licensed under the Apache-2.0 license
 
-pub mod cert_store;
-mod device_cert;
-mod endorsement_cert;
-mod leaf_cert;
+pub(crate) mod cert_store;
+pub(crate) mod device_cert;
+pub(crate) mod endorsement_cert;
+pub(crate) mod leaf_cert;
